+++ packages/core/.storybook/main.js
@@ -26,10 +26,25 @@ module.exports = {
       }),
     ]
 
-    // SCSS import support
+    // SCSS import and camelCase support for CSS modules
     config.module.rules.push({
       test: /\.scss$/,
-      use: ['style-loader', 'css-loader', 'sass-loader'],
+      use: [
+        {
+          loader: 'style-loader',
+        },
+        {
+          loader: 'css-loader',
+          options: {
+            modules: {
+              exportLocalsConvention: 'camelCase',
+            },
+          },
+        },
+        {
+          loader: 'sass-loader',
+        },
+      ],
       include: resolve(__dirname, '../'),
     })